Experience auth

+++ app/controllers/api/experiences_controller.rb
@@ -1,4 +1,7 @@
 class Api::ExperiencesController < ApplicationController
+
+  before_action :authenticate_student, only: [:create, :update, :destroy]
+  
   def index
     @experiences = Experience.all
     render "index.json.jb"
@@ -27,22 +30,30 @@ class Api::ExperiencesController < ApplicationController
 
   def update
     @experience = Experience.find_by(id: params[:id])
+    if current_student.id == @experience.student_id
 
-    @experience.start_date = params[:start_date] || @experience.start_date
-    @experience.end_date = params[:end_date] || @experience.end_date
-    @experience.job_title = params[:job_title] || @experience.job_title
-    @experience.company = params[:company] || @experience.company
-    @experience.details = params[:details] || @experience.details
-    
+      @experience.start_date = params[:start_date] || @experience.start_date
+      @experience.end_date = params[:end_date] || @experience.end_date
+      @experience.job_title = params[:job_title] || @experience.job_title
+      @experience.company = params[:company] || @experience.company
+      @experience.details = params[:details] || @experience.details
+      
 
-    @experience.save
-    render "show.json.jb" 
+      @experience.save
+      render "show.json.jb" 
+    else
+      render json: {message: "Experience can only be edited by original owner. "}
+    end  
   end
 
   def destroy
     experience = Experience.find_by(id: params[:id])
-    experience.destroy
-    render json: { message: "Your work experience has been removed from the database" }
+    if current_student.id == experience.student_id
+      experience.destroy
+      render json: { message: "Your work experience has been removed from the database" }
+    else
+      render json: {error: "Experience can only be deleted by original owner."}, status: :unauthorized
+    end   
     
   end
 end
